+++ lib/index.js
@@ -4,22 +4,22 @@ const BaseFileModel = require('./base-model');
 const SlsApiUpload = require('./sls-api-upload');
 const SlsApiFileRelation = require('./sls-api-file-relation');
 const SlsApiFileDelete = require('./sls-api-file-delete');
-const SlsApiFileDownload = require('./sls-api-file-download');
+const SlsApiFileGet = require('./sls-api-file-get');
 const SlsApiFileList = require('./sls-api-file-list');
 const SlsApiFileRelationError = require('./sls-api-file-relation-error');
 const SlsApiUploadError = require('./sls-api-upload-error');
 const SlsApiFileDeleteError = require('./sls-api-file-delete-error');
-const SlsApiFileDownloadError = require('./sls-api-file-download-error');
+const SlsApiFileGetError = require('./sls-api-file-get-error');
 
 module.exports = {
 	BaseFileModel,
 	SlsApiUpload,
 	SlsApiFileRelation,
 	SlsApiFileDelete,
-	SlsApiFileDownload,
+	SlsApiFileGet,
 	SlsApiFileList,
 	SlsApiFileRelationError,
 	SlsApiUploadError,
 	SlsApiFileDeleteError,
-	SlsApiFileDownloadError
+	SlsApiFileGetError
 };
