+++ lib/sls-api-file-get-error.js
@@ -1,6 +1,6 @@
 'use strict';
 
-class SlsApiFileDownloadError extends Error {
+class SlsApiFileGetError extends Error {
 
 	static get messages() {
 
@@ -19,8 +19,8 @@ class SlsApiFileDownloadError extends Error {
 	constructor(err) {
 		super(err);
 		this.message = err.message || err;
-		this.name = 'SlsApiFileDownloadError';
+		this.name = 'SlsApiFileGetError';
 	}
 }
 
-module.exports = SlsApiFileDownloadError;
+module.exports = SlsApiFileGetError;
